requirements.txt
asgiref==3.7.2
autopep8==2.0.4
bcrypt==4.0.1
blinker==1.6.2
boto3==1.28.54
botocore==1.31.54
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.4
defusedxml==0.7.1
distlib==0.3.7
dj-database-url==2.1.0
Django==4.2.5
django-cors-headers==4.2.0
django-ses==3.5.0
django-storages==1.14
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
filelock==3.12.4
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jovian==0.2.47
MarkupSafe==2.1.3
oauthlib==3.2.2
Pillow==10.0.0
platformdirs==3.10.0
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.2
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
SQLAlchemy==2.0.20
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==1.26.16
uuid==1.30
virtualenv==20.24.5
Werkzeug==2.3.6
